A minor change to conform with -Werror
The compilation of a program which uses libcsv will halt if -Werror is specified due to the variable 'pstate', which is being set but not used. 
$ warning: variable 'pstate' set but not used [-Wunused-but-set-variable]

I can see that 'pstate' is being declared for consistency with the shared macros, but that it doesn't otherwise deviate from the value of 'p->pstate' before the switch (they were equal throughout the tests). So I'm proposing a simple change to give 'pstate' a job.

I might be totally wrong about this, so don't hurt me. I left a note in my pocket naming you as my killer.